settings and disclaimer
